requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.8
coverage==7.8.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
flake8==7.2.0
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mirakuru==2.6.0
multidict==6.4.3
mypy==1.15.0
mypy-extensions==1.0.0
packaging==25.0
pluggy==1.5.0
port-for==0.7.4
propcache==0.3.1
psutil==7.0.0
psycopg==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-postgresql==7.0.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
types-pyasn1==0.6.0.20250208
types-python-jose==3.4.0.20250224
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
yarl==1.20.0
